[CHORE] #35 : module rename navigation -> common
