+++ android/app/src/main/java/com/jaino/app/di/AppModule.kt
@@ -2,7 +2,7 @@ package com.jaino.app.di
 
 import android.app.Application
 import com.jaino.app.navigator.AppNavigatorImpl
-import com.jaino.navigation.AppNavigator
+import com.jaino.common.AppNavigator
 import com.kakao.sdk.user.UserApiClient
 import dagger.Module
 import dagger.Provides
