+++ android/core/common/.gitignore
