+++ android/core/common/proguard-rules.pro
