+++ android/core/common/src/androidTest/java/com/jaino/common/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.jaino.navigation
+package com.jaino.common
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
