+++ android/core/common/src/main/AndroidManifest.xml
