+++ android/feature/analyze/build.gradle.kts
@@ -16,7 +16,7 @@ dependencies {
     implementation(project(":core:data"))
     implementation(project(":core:domain"))
     implementation(project(":core:model"))
-    implementation(project(":core:navigation"))
+    implementation(project(":core:common"))
     implementation(project(":core:designsystem"))
     implementation(libs.bundles.camerax)
     implementation(libs.timber)
