+++ android/settings.gradle
@@ -30,7 +30,7 @@ include ':core:model'
 include ':core:domain'
 include ':core:network'
 include ':core:datastore'
-include ':core:navigation'
+include ':core:common'
 include ':feature:auth'
 include ':feature:setting'
 include ':feature:analyze'
